foundry_agent_backend_api/requirements.txt
fastapi
uvicorn
pyjwt
cryptography
requests
azure-ai-projects>=2.0.0b1
python-dotenv